Deprecate usage of XML Spring schemas
Move Template test

Related to #141.
